packet queueing fixes
