+++ src/stack-ndpv6.h
@@ -2,7 +2,7 @@
 #define STACK_NDPV6_H
 #include <stddef.h>
 #include <time.h>
-#include "packet-queue.h"
+#include "stack-queue.h"
 #include "ipv6address.h"
 struct PreprocessedInfo;
 
@@ -12,6 +12,9 @@ stack_handle_ndpv6_neighbor_notification(
             const unsigned char *buf, size_t length,
             const unsigned char *mac_them,
             struct stack_t *stack);
+
+int
+stack_handle_neighbor_solicitation(struct stack_t *stack, struct PreprocessedInfo *parsed,  const unsigned char *px, size_t length);
 
 #endif
